fix(cart&cartProduct): change DECIMAL type to Float at models

See merge request gustavo.hmessias96/kenzie-shop!22

+++ app/models/cart_model.py
@@ -1,5 +1,5 @@
 from app.config.database import db
-from sqlalchemy import Column, Integer, Boolean, DECIMAL
+from sqlalchemy import Column, Integer, Boolean, Float
 from sqlalchemy.orm import relationship, backref
 
 from dataclasses import dataclass
@@ -16,7 +16,7 @@ class CartModel(db.Model):
     id = Column(Integer, primary_key=True)
 
     is_empty = Column(Boolean, nullable=False)
-    total_price = Column(DECIMAL, nullable=False)
+    total_price = Column(Float, nullable=False)
 
     products = relationship(
         "ProductModel", secondary="carts_products", backref=backref("carts")
